feat: Added to WCS serial communication and motor control

+++ WCS/include/Slider.h
@@ -7,7 +7,7 @@ class Slider {
 
     public:
         virtual bool isChanged() = 0;
-        virtual uint16_t getValue() = 0;
+        virtual uint8_t getValue() = 0; // Returns a value in percentage [0, 100]
         virtual void start() = 0;
 
 };
